+++ NEWS.md
@@ -1,2 +1,7 @@
+0.2.0
+* allow passing through of params to httr::HEAD() via ...
+* as.data.frame method for httr::headers() return (dispatches on
+  objects classed 'invisible')
+
 0.1.0 
 * Initial release